requirements.txt
# Web framework
flask>=3.0.0
gunicorn>=21.0.0

# For production WSGI server
gevent>=24.0.0

# Telnyx SDK for SMS and API calls
telnyx>=2.0.0

# Environment variable loading
python-dotenv>=1.0.0
